Suhani_Bansal/requirements.txt
aiohttp
beautifulsoup4
trafilatura
numpy
sentence-transformers
faiss-cpu
requests